Revert "Revert "[objc] Changing GRPCErrorCode enum base type to int32_t (#27908)" (#27921)"
This reverts commit 596b919707004c90540cba0eb7b337a323f27953.

+++ src/objective-c/GRPCClient/GRPCTypes.h
@@ -23,7 +23,7 @@
  * Note that a few of these are never produced by the gRPC libraries, but are of
  * general utility for server applications to produce.
  */
-typedef NS_ENUM(NSUInteger, GRPCErrorCode) {
+typedef NS_ENUM(int32_t, GRPCErrorCode) {
   /** The operation was cancelled (typically by the caller). */
   GRPCErrorCodeCancelled = 1,
